Can now compile in libintense-io and libintense. 	modified:   libintense-io/Makefile.in 	modified:   libintense-io/autom4te.cache/requests 	modified:   libintense-io/config/install-sh 	modified:   libintense-io/config/missing 	modified:   libintense-io/include/intense-io/BaseSerialiser.hpp 	modified:   libintense-io/include/intense-io/BinaryBaseSerialiser.hpp 	modified:   libintense-io/include/intense-io/LogStream.hpp 	deleted:    libintense-io/install-sh 	deleted:    libintense-io/missing 	modified:   libintense-io/src/BinaryBaseSerialiser.cpp 	modified:   libintense-io/tests/BinarySerialiserTestSource.cpp 	modified:   libintense/Makefile.am 	modified:   libintense/Makefile.in 	modified:   libintense/aclocal.m4 	modified:   libintense/autom4te.cache/output.0 	modified:   libintense/autom4te.cache/output.1 	new file:   libintense/autom4te.cache/output.2 	modified:   libintense/autom4te.cache/requests 	modified:   libintense/autom4te.cache/traces.0 	modified:   libintense/autom4te.cache/traces.1 	new file:   libintense/autom4te.cache/traces.2 	modified:   libintense/config/install-sh 	modified:   libintense/config/ltmain.sh 	modified:   libintense/config/missing 	new file:   libintense/config/test-driver 	modified:   libintense/configure 	modified:   libintense/configure.ac 	modified:   libintense/include/Makefile.in 	deleted:    libintense/install-sh 	new file:   libintense/m4/libtool.m4 	new file:   libintense/m4/ltoptions.m4 	new file:   libintense/m4/ltsugar.m4 	new file:   libintense/m4/ltversion.m4 	new file:   libintense/m4/lt~obsolete.m4 	deleted:    libintense/missing 	modified:   libintense/src/Makefile.am 	modified:   libintense/src/Makefile.in 	modified:   libintense/tests/Makefile.am 	modified:   libintense/tests/Makefile.in

+++ libintense-io/include/intense-io/BaseSerialiser.hpp
@@ -130,16 +130,16 @@ namespace intense {
       (short unsigned int& i, std::istream& is) = 0;
 
       virtual void out
-      (long long int i, std::ostream& os) = 0;
+      (long int i, std::ostream& os) = 0;
 
       virtual void in
-      (long long int& i, std::istream& is) = 0;
+      (long int& i, std::istream& is) = 0;
 
       virtual void out
-      (unsigned long long int i, std::ostream& os) = 0;
+      (unsigned long int i, std::ostream& os) = 0;
 
       virtual void in
-      (unsigned long long int& i, std::istream& is) = 0;
+      (unsigned long int& i, std::istream& is) = 0;
 
       virtual void out
       (char c, std::ostream& os) = 0;
